refactor(Workshop): rename file
